Data.Dense

Description

This module provides a large subset of the full functionality of "dense" without exporting names that conflict with names in prelude, so it can often be imported unqualified. It also includes reexported classes and data types from other modules. However it does not contain much functions necessary to construct arrays, for that see Data.Dense.Generic or one of the type specific modules intended to be imported qualified. Typical imports for shaped will look like this:

import           Data.Dense
import qualified Data.Dense.Unboxed as U

For boxed-specific arrays (a la Data.Vector) see Data.Dense.Boxed.

Array types

data Array v f a

An Array is a vector with a shape.

type BArray = Array Vector

Boxed array.

type UArray = Array Vector

Unboxed array.

type SArray = Array Vector

Storeable array.

type PArray = Array Vector

Prim array.

Indexing

type Layout f = f Int

A Layout is the full size of an array. This alias is used to help distinguish between the layout of an array and an index (usually just l Int) in a type signature.

class Shape f => HasLayout f a | a -> f where

Class of things that have a Layout. This means we can use the same functions for the various different arrays in the library.

Methods

layout :: Lens' a (Layout f)

Lens onto the Layout of something.

layout :: a ~ f Int => (Layout f -> g (Layout f)) -> a -> g a

Lens onto the Layout of something.

class (Eq1 f, Additive f, Traversable f) => Shape f

Class for types that can be converted to and from linear indexes.

extent :: HasLayout f a => a -> f Int

Get the extent of an array.

extent :: Array v f a    -> f Int
extent :: MArray v f s a -> f Int
extent :: Delayed f a    -> f Int
extent :: Focused f a    -> f Int

size :: HasLayout f a => a -> Int

Get the total number of elements in an array.

size :: Array v f a    -> Int
size :: MArray v f s a -> Int
size :: Delayed f a    -> Int
size :: Focused f a    -> Int

Folds over indexes

indexes :: HasLayout f a => IndexedFold Int a (f Int)

Indexed fold for all the indexes in the layout.

indexesBetween :: HasLayout f a => f Int -> f Int -> IndexedFold Int a (f Int)

Indexed fold between the two indexes where the index is the linear index for the original layout.

indexesFrom :: HasLayout f a => f Int -> IndexedFold Int a (f Int)

Indexed fold starting starting from some point, where the index is the linear index for the original layout.

Lenses

vector :: (Vector v a, Vector w b) => IndexedLens (Layout f) (Array v f a) (Array w f b) (v a) (w b)

Indexed lens over the underlying vector of an array. The index is the extent of the array. You must _not_ change the length of the vector, otherwise an error will be thrown (even for V1 layouts, use flat for V1).

Traversals

values :: (Shape f, Vector v a, Vector w b) => IndexedTraversal (f Int) (Array v f a) (Array w f b) a b

Indexed traversal over the elements of an array. The index is the current position in the array.

values' :: (Shape f, Vector v a, Vector v b) => IndexedTraversal (f Int) (Array v f a) (Array v f b) a b

Same as values but restrictive in the vector type.

valuesBetween :: (Shape f, Vector v a) => f Int -> f Int -> IndexedTraversal' (f Int) (Array v f a) a

Traverse over the values between two indexes.

Construction

Flat arrays

flat :: Vector w b => Iso (Array v V1 a) (Array w V1 b) (v a) (w b)

1D arrays are just vectors. You are free to change the length of the vector when going over this Iso (unlike linear).

Note that V1 arrays are an instance of Vector so you can use any of the functions in Data.Vector.Generic on them without needing to convert.

Shaped from lists

fromListInto :: (Shape f, Vector v a) => Layout f -> [a] -> Maybe (Array v f a)

O(n) Convert the first n elements of a list to an Array with the given shape. Returns Nothing if there are not enough elements in the list.

fromListInto_ :: (Shape f, Vector v a) => Layout f -> [a] -> Array v f a

O(n) Convert the first n elements of a list to an Array with the given shape. Throw an error if the list is not long enough.

Shaped from vectors

fromVectorInto :: (Shape f, Vector v a) => Layout f -> v a -> Maybe (Array v f a)

Create an array from a vector and a layout. Return Nothing if the vector is not the right shape.

fromVectorInto_ :: (Shape f, Vector v a) => Layout f -> v a -> Array v f a

Create an array from a vector and a layout. Throws an error if the vector is not the right shape.

Generating

See Data.Shaped.Generic.

Functions on arrays

Empty arrays

See AsEmpty class or Data.Shaped.Generic.

Indexing

See Ixed class.

Modifying arrays

See Data.Shaped.Generic.

Slices

Matrix

ixRow :: Vector v a => Int -> IndexedTraversal' Int (Array v V2 a) (v a)

Affine traversal over a single row in a matrix.

>>> traverseOf_ rows print $ m & ixRow 1 . each *~ 2
[a,b,c,d]
[e * 2,f * 2,g * 2,h * 2]
[i,j,k,l]

The row vector should remain the same size to satisfy traversal laws but give reasonable behaviour if the size differs:

>>> traverseOf_ rows print $ m & ixRow 1 .~ B.fromList [0,1]
[a,b,c,d]
[0,1,g,h]
[i,j,k,l]

>>> traverseOf_ rows print $ m & ixRow 1 .~ B.fromList [0..100]
[a,b,c,d]
[0,1,2,3]
[i,j,k,l]

rows :: (Vector v a, Vector w b) => IndexedTraversal Int (Array v V2 a) (Array w V2 b) (v a) (w b)

Indexed traversal over the rows of a matrix. Each row is an efficient slice of the original vector.

>>> traverseOf_ rows print m
[a,b,c,d]
[e,f,g,h]
[i,j,k,l]

ixColumn :: Vector v a => Int -> IndexedTraversal' Int (Array v V2 a) (v a)

Affine traversal over a single column in a matrix.

>>> traverseOf_ rows print $ m & ixColumn 2 . each +~ 1
[a,b,c + 1,d]
[e,f,g + 1,h]
[i,j,k + 1,l]

columns :: (Vector v a, Vector w b) => IndexedTraversal Int (Array v V2 a) (Array w V2 b) (v a) (w b)

Indexed traversal over the columns of a matrix. Unlike rows, each column is a new separate vector.

>>> traverseOf_ columns print m
[a,e,i]
[b,f,j]
[c,g,k]
[d,h,l]

>>> traverseOf_ rows print $ m & columns . indices odd . each .~ 0
[a,0,c,0]
[e,0,g,0]
[i,0,k,0]

The vectors should be the same size to be a valid traversal. If the vectors are different sizes, the number of rows in the new array will be the length of the smallest vector.

3D

ixPlane :: Vector v a => ALens' (V3 Int) (V2 Int) -> Int -> IndexedTraversal' Int (Array v V3 a) (Array v V2 a)

Traversal over a single plane of a 3D array given a lens onto that plane (like _xy, _yz, _zx).

planes :: (Vector v a, Vector w b) => ALens' (V3 Int) (V2 Int) -> IndexedTraversal Int (Array v V3 a) (Array w V3 b) (Array v V2 a) (Array w V2 b)

Traversal over all planes of 3D array given a lens onto that plane (like _xy, _yz, _zx).

flattenPlane :: (Vector v a, Vector w b) => ALens' (V3 Int) (V2 Int) -> (v a -> b) -> Array v V3 a -> Array w V2 b

Flatten a plane by reducing a vector in the third dimension to a single value.

Mutable

data MArray v l s a

A mutable array with a shape.

type BMArray = MArray MVector

Boxed mutable array.

type UMArray = MArray MVector

Unboxed mutable array.

type SMArray = MArray MVector

Storable mutable array.

type PMArray = MArray MVector

Primitive mutable array.

Delayed

data Delayed f a

A delayed representation of an array. This useful for mapping over an array in parallel.

Generating delayed

delayed :: (Vector v a, Vector w b, Shape f, Shape g) => Iso (Array v f a) (Array w g b) (Delayed f a) (Delayed g b)

Isomorphism between an array and its delayed representation. Conversion to the array is done in parallel.

seqDelayed :: (Vector v a, Vector w b, Shape f, Shape g) => Iso (Array v f a) (Array w g b) (Delayed f a) (Delayed g b)

Isomorphism between an array and its delayed representation. Conversion to the array is done in parallel.

delay :: (Vector v a, Shape f) => Array v f a -> Delayed f a

Turn a material array into a delayed one with the same shape.

manifest :: (Vector v a, Shape f) => Delayed f a -> Array v f a

Parallel manifestation of a delayed array into a material one.

seqManifest :: (Vector v a, Shape f) => Delayed f a -> Array v f a

Sequential manifestation of a delayed array.

genDelayed :: Layout f -> (f Int -> a) -> Delayed f a

Generate a Delayed array using the given Layout and construction function.

indexDelayed :: Shape f => Delayed f a -> f Int -> a

Index a delayed array, returning a IndexOutOfBounds exception if the index is out of range.

affirm :: (Shape f, Unbox a) => Delayed f a -> Delayed f a

manifest an array to a UArray and delay again. See Data.Dense.Boxed or Data.Dense.Storable to affirm for other types of arrays.

seqAffirm :: (Shape f, Unbox a) => Delayed f a -> Delayed f a

seqManifest an array to a UArray and delay again. See Data.Dense.Boxed or Data.Dense.Storable to affirm for other types of arrays.

Helpful reexports

(*^) :: (Functor f, Num a) => a -> f a -> f a infixl 7

Compute the left scalar product

>>> 2 *^ V2 3 4
V2 6 8

(^*) :: (Functor f, Num a) => f a -> a -> f a infixl 7

Compute the right scalar product

>>> V2 3 4 ^* 2
V2 6 8

(^/) :: (Functor f, Fractional a) => f a -> a -> f a infixl 7

Compute division by a scalar on the right.

class Functor f => Additive f where

A vector is an additive group with additional structure.

Minimal complete definition

Nothing

class Additive f => Metric f where

Free and sparse inner product/metric spaces.

Minimal complete definition

Nothing

Focused

data Focused f a

A delayed representation of an array with a focus on a single element. This element is the target of extract.

Generating focused

focusOn :: f Int -> Delayed f a -> Focused f a

Focus on a particular element of a delayed array.

unfocus :: Focused f a -> Delayed f a

Discard the focus to retrieve the delayed array.

unfocused :: IndexedLens (f Int) (Focused f a) (Focused f b) (Delayed f a) (Delayed f b)

Indexed lens onto the delayed array, indexed at the focus.

extendFocus :: Shape f => (Focused f a -> b) -> Delayed f a -> Delayed f b

Modify a Delayed array by extracting a value from a Focused each point.

Focus location

locale :: ComonadStore s w => Lens' (w a) s

Lens onto the position of a ComonadStore.

locale :: Lens' (Focused l a) (l Int)

shiftFocus :: Applicative f => f Int -> Focused f a -> Focused f a

Focus on a neighbouring element, relative to the current focus.

Boundary

data Boundary

The boundary condition used for indexing relative elements in a Focused.

Constructors

Clamp
clamp coordinates to the extent of the array

Mirror
mirror coordinates beyond the array extent

Wrap
wrap coordinates around on each dimension

peekB :: Shape f => Boundary -> f Int -> Focused f a -> a

Index a focused using a Boundary condition.

peeksB :: Shape f => Boundary -> (f Int -> f Int) -> Focused f a -> a

Index an element by applying a function the current position, using a boundary condition.

peekRelativeB :: Shape f => Boundary -> f Int -> Focused f a -> a

Index an element relative to the current focus using a Boundary condition.

Helpful reexports

class Functor w => Comonad w where

There are two ways to define a comonad:

I. Provide definitions for extract and extend satisfying these laws:

extend extract      = id
extract . extend f  = f
extend f . extend g = extend (f . extend g)

In this case, you may simply set fmap = liftW.

These laws are directly analogous to the laws for monads and perhaps can be made clearer by viewing them as laws stating that Cokleisli composition must be associative, and has extract for a unit:

f =>= extract   = f
extract =>= f   = f
(f =>= g) =>= h = f =>= (g =>= h)

II. Alternately, you may choose to provide definitions for fmap, extract, and duplicate satisfying these laws:

extract . duplicate      = id
fmap extract . duplicate = id
duplicate . duplicate    = fmap duplicate . duplicate

In this case you may not rely on the ability to define fmap in terms of liftW.

You may of course, choose to define both duplicate and extend. In that case you must also satisfy these laws:

extend f  = fmap f . duplicate
duplicate = extend id
fmap f    = extend (f . extract)

These are the default definitions of extend and duplicate and the definition of liftW respectively.

Minimal complete definition

extract, (duplicate | extend)

class Comonad w => ComonadStore s w | w -> s where

Minimal complete definition

pos, peek

Stencils

data Stencil f a

Stencils are used to fold over neighbouring array sites. To construct a stencil use mkStencil, mkStencilUnboxed. For static sized stencils you can use the quasiquoter stencil.

To use a stencil you can use stencilSum or use the Foldable and FoldableWithIndex instances.

Constructing stencils

stencil :: QuasiQuoter

QuasiQuoter for producing a static stencil definition. This is a versatile parser for 1D, 2D and 3D stencils. The parsing is similar to dense but stencil also supports _, which means ignore this element. Also, stencils should have an odd length in all dimensions so there is always a center element (which is used as zero).

Examples

1D stencils are of the form

[stencil| 5 -3 1 -3 5 |] :: Num a => Stencil V1 a

2D stencils are of the form

myStencil2 :: Num a => Stencil V2 a
myStencil2 = [stencil|
  0 1 0
  1 0 1
  0 1 0
|]

3D stencils have gaps between planes.

myStencil3 :: Fractional a => Stencil V3 a
myStencil3 :: [stencil|
  1/20 3/10 1/20
  3/10  1   3/10
  1/20 3/10 1/20

  3/10  1   3/10
   1    _    1
  3/10  1   3/10

  1/20 3/10 1/20
  3/10  1   3/10
  1/20 3/10 1/20
|]

Variables can also be used

myStencil2' :: a -> a -> a -> Stencil V2 a
myStencil2' a b c = [stencil|
  c b c
  b a b
  c b c
|]

mkStencil :: [(f Int, a)] -> Stencil f a

Make a stencil folding over a list.

If the list is staticlly known this should expand at compile time via rewrite rules, similar to makeStencilTH but less reliable. If that does not happen the resulting could be slow. If the list is not know at compile time, mkStencilUnboxed can be signifcantly faster (but isn't subject expending via rewrite rules).

mkStencilTH :: (ShapeLift f, Lift a) => [(f Int, a)] -> Q Exp

Construct a Stencil by unrolling the list at compile time. For example

ifoldr f b $(mkStencilTH [(V1 (-1), 5), (V1 0, 3), (V1 1, 5)])

will be get turned into

f (V1 (-1)) 5 (f (V1 0) 3 (f (V1 1) 5 b))

at compile time. Since there are no loops and all target indexes are known at compile time, this can lead to more optimisations and faster execution times. This can lead to around a 2x speed up compared to folding over unboxed vectors.

myStencil = $(mkStencilTH (as :: [(f Int, a)])) :: Stencil f a

Using stencils

stencilSum :: (Shape f, Num a) => Boundary -> Stencil f a -> Focused f a -> a

Sum the elements around a Focused using a Boundary condition and a Stencil.

This is often used in conjunction with extendFocus.

Common shapes

newtype V1 a :: * -> *

A 1-dimensional vector

>>> pure 1 :: V1 Int
V1 1

>>> V1 2 + V1 3
V1 5

>>> V1 2 * V1 3
V1 6

>>> sum (V1 2)
2

Constructors

V1 a

data V2 a :: * -> *

A 2-dimensional vector

>>> pure 1 :: V2 Int
V2 1 1

>>> V2 1 2 + V2 3 4
V2 4 6

>>> V2 1 2 * V2 3 4
V2 3 8

>>> sum (V2 1 2)
3

Constructors

V2 ~a ~a

data V3 a :: * -> *

A 3-dimensional vector

Constructors

V3 ~a ~a ~a

data V4 a :: * -> *

A 4-dimensional vector.

Constructors

V4 ~a ~a ~a ~a

class R1 t where

A space that has at least 1 basis vector _x.

Minimal complete definition

Nothing

class R1 t => R2 t where

A space that distinguishes 2 orthogonal basis vectors _x and _y, but may have more.

Minimal complete definition

Nothing

class R2 t => R3 t where

A space that distinguishes 3 orthogonal basis vectors: _x, _y, and _z. (It may have more)

Minimal complete definition

Nothing

class R3 t => R4 t where

A space that distinguishes orthogonal basis vectors _x, _y, _z, _w. (It may have more.)

Minimal complete definition

Nothing

Extra planes

_xz :: R3 t => Lens' (t a) (V2 a)

_yz :: R3 t => Lens' (t a) (V2 a)

_yx :: R2 t => Lens' (t a) (V2 a)

>>> V2 1 2 ^. _yx
V2 2 1

_zy :: R3 t => Lens' (t a) (V2 a)

_zx :: R3 t => Lens' (t a) (V2 a)