Data.Dense.Mutable

Description

This module provides generic functions over mutable multidimensional arrays.

Mutable array

data MArray v l s a

A mutable array with a shape.

Constructors

MArray !(Layout l) !(v s a)

type UMArray = MArray MVector

Unboxed mutable array.

type SMArray = MArray MVector

Storable mutable array.

type BMArray = MArray MVector

Boxed mutable array.

type PMArray = MArray MVector

Primitive mutable array.

Lenses

mlayout :: (Shape l, Shape l') => Lens (MArray v l s a) (MArray v l' s a) (Layout l) (Layout l')

Lens onto the shape of the vector. The total size of the layout _must_ remain the same or an error is thrown.

mvector :: (MVector v a, MVector w b) => IndexedLens (Layout l) (MArray v l s a) (MArray w l t b) (v s a) (w t b)

Indexed lens over the underlying vector of an array. The index is the extent of the array. You must not change the length of the vector, otherwise an error will be thrown.

Creation

new :: (PrimMonad m, Shape l, MVector v a) => Layout l -> m (MArray v l (PrimState m) a)

New mutable array with shape l.

replicate :: (PrimMonad m, Shape l, MVector v a) => Layout l -> a -> m (MArray v l (PrimState m) a)

New mutable array with shape l filled with element a.

replicateM :: (PrimMonad m, Shape l, MVector v a) => Layout l -> m a -> m (MArray v l (PrimState m) a)

New mutable array with shape l filled with result of monadic action a.

clone :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> m (MArray v l (PrimState m) a)

Clone a mutable array, making a new, separate mutable array.

Standard operations

Indexing

read :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> m a

Read a mutable array at element l.

linearRead :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> m a

Read a mutable array at element i by indexing the internal vector.

unsafeRead :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> m a

read without bounds checking.

unsafeLinearRead :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> m a

linearRead without bounds checking.

Writing

write :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> a -> m ()

Write a mutable array at element l.

linearWrite :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> a -> m ()

Write a mutable array at element i by indexing the internal vector.

unsafeWrite :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> a -> m ()

write without bounds checking.

unsafeLinearWrite :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> a -> m ()

linearWrite without bounds checking.

Modifying

modify :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> (a -> a) -> m ()

Modify a mutable array at element l by applying a function.

linearModify :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> (a -> a) -> m ()

Modify a mutable array at element i by applying a function.

unsafeModify :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> (a -> a) -> m ()

modify without bounds checking.

unsafeLinearModify :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> (a -> a) -> m ()

linearModify without bounds checking.

Swap

swap :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> l Int -> m ()

Swap two elements in a mutable array.

linearSwap :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> Int -> m ()

Swap two elements in a mutable array by indexing the internal vector.

unsafeSwap :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> l Int -> m ()

swap without bounds checking.

unsafeLinearSwap :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> Int -> m ()

linearSwap without bounds checking.

Exchange

exchange :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> a -> m a

Replace the element at the give position and return the old element.

linearExchange :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> a -> m a

Replace the element at the give position and return the old element.

unsafeExchange :: (PrimMonad m, Shape l, MVector v a) => MArray v l (PrimState m) a -> l Int -> a -> m a

Replace the element at the give position and return the old element.

unsafeLinearExchange :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> Int -> a -> m a

Replace the element at the give position and return the old element.

Misc

set :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> a -> m ()

Set all elements in a mutable array to a constant value.

clear :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> m ()

Clear the elements of a mutable array. This is usually a no-op for unboxed arrays.

copy :: (PrimMonad m, MVector v a) => MArray v l (PrimState m) a -> MArray v l (PrimState m) a -> m ()

Copy all elements from one array into another.