Data.Dense.Storable

Description

Storeable multidimentional arrays.

SArray types

type SArray = Array Vector

Storeable array.

class Storable a

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)

class (Eq1 f, Additive f, Traversable f) => Shape f

Class for types that can be converted to and from linear indexes.

Layout of an array

class Shape f => HasLayout f a | a -> f where

Class of things that have a Layout. This means we can use the same functions for the various different arrays in the library.

Methods

layout :: Lens' a (Layout f)

Lens onto the Layout of something.

layout :: a ~ f Int => (Layout f -> g (Layout f)) -> a -> g a

Lens onto the Layout of something.

type Layout f = f Int

A Layout is the full size of an array. This alias is used to help distinguish between the layout of an array and an index (usually just l Int) in a type signature.

Extracting size

extent :: HasLayout f a => a -> f Int

Get the extent of an array.

extent :: Array v f a    -> f Int
extent :: MArray v f s a -> f Int
extent :: Delayed f a    -> f Int
extent :: Focused f a    -> f Int

size :: HasLayout f a => a -> Int

Get the total number of elements in an array.

size :: Array v f a    -> Int
size :: MArray v f s a -> Int
size :: Delayed f a    -> Int
size :: Focused f a    -> Int

Folds over indexes

indexes :: HasLayout f a => IndexedFold Int a (f Int)

Indexed fold for all the indexes in the layout.

indexesFrom :: HasLayout f a => f Int -> IndexedFold Int a (f Int)

Indexed fold starting starting from some point, where the index is the linear index for the original layout.

indexesBetween :: HasLayout f a => f Int -> f Int -> IndexedFold Int a (f Int)

Indexed fold between the two indexes where the index is the linear index for the original layout.

Underlying vector

vector :: (Storable a, Storable b) => IndexedLens (Layout f) (SArray f a) (SArray f b) (Vector a) (Vector b)

Indexed lens over the underlying vector of an array. The index is the extent of the array. You must _not_ change the length of the vector, otherwise an error will be thrown (even for V1 layouts, use flat for V1).

Traversals

values :: (Shape f, Storable a, Storable b) => IndexedTraversal (f Int) (SArray f a) (SArray f b) a b

Same as values but restrictive in the vector type.

values' :: (Shape f, Storable a, Storable b) => IndexedTraversal (f Int) (SArray f a) (SArray f b) a b

Same as values but restrictive in the vector type.

valuesBetween :: (Shape f, Storable a) => f Int -> f Int -> IndexedTraversal' (f Int) (SArray f a) a

Same as values but restrictive in the vector type.

Construction

Flat arrays

flat :: Storable b => Iso (SArray V1 a) (SArray V1 b) (Vector a) (Vector b)

1D arrays are just vectors. You are free to change the length of the vector when going over this Iso (unlike linear).

Note that V1 arrays are an instance of Vector so you can use any of the functions in Generic on them without needing to convert.

fromList :: Storable a => [a] -> SArray V1 a

Contruct a flat array from a list. (This is just fromList from Generic.)

From lists

fromListInto :: (Shape f, Storable a) => Layout f -> [a] -> Maybe (SArray f a)

O(n) Convert the first n elements of a list to an SArrayith the given shape. Returns Nothing if there are not enough elements in the list.

fromListInto_ :: (Shape f, Storable a) => Layout f -> [a] -> SArray f a

O(n) Convert the first n elements of a list to an SArrayith the given shape. Throw an error if the list is not long enough.

From vectors

fromVectorInto :: (Shape f, Storable a) => Layout f -> Vector a -> Maybe (SArray f a)

Create an array from a vector and a layout. Return Nothing if the vector is not the right shape.

fromVectorInto_ :: (Shape f, Storable a) => Layout f -> Vector a -> SArray f a

Create an array from a vector and a layout. Throws an error if the vector is not the right shape.

Initialisation

replicate :: (Shape f, Storable a) => f Int -> a -> SArray f a

O(n) SArray of the given shape with the same value in each position.

generate :: (Shape f, Storable a) => Layout f -> (f Int -> a) -> SArray f a

O(n) Construct an array of the given shape by applying the function to each index.

linearGenerate :: (Shape f, Storable a) => Layout f -> (Int -> a) -> SArray f a

O(n) Construct an array of the given shape by applying the function to each index.

Monadic initialisation

create :: Storable a => (forall s. ST s (SMArray f s a)) -> SArray f a

Execute the monadic action and freeze the resulting array.

replicateM :: (Monad m, Shape f, Storable a) => Layout f -> m a -> m (SArray f a)

O(n) Construct an array of the given shape by filling each position with the monadic value.

generateM :: (Monad m, Shape f, Storable a) => Layout f -> (f Int -> m a) -> m (SArray f a)

O(n) Construct an array of the given shape by applying the monadic function to each index.

linearGenerateM :: (Monad m, Shape f, Storable a) => Layout f -> (Int -> m a) -> m (SArray f a)

O(n) Construct an array of the given shape by applying the monadic function to each index.

Functions on arrays

Empty arrays

empty :: (Storable a, Additive f) => SArray f a

The empty SArray with a zero shape.

null :: Foldable f => SArray f a -> Bool

Test is if the array is empty.

Indexing

(!) :: (Shape f, Storable a) => SArray f a -> f Int -> a

Index an element of an array. Throws IndexOutOfBounds if the index is out of bounds.

(!?) :: (Shape f, Storable a) => SArray f a -> f Int -> Maybe a

Safe index of an element.

unsafeIndex :: (Shape f, Storable a) => SArray f a -> f Int -> a

Index an element of an array without bounds checking.

linearIndex :: Storable a => SArray f a -> Int -> a

Index an element of an array while ignoring its shape.

unsafeLinearIndex :: Storable a => SArray f a -> Int -> a

Index an element of an array while ignoring its shape, without bounds checking.

Monadic indexing

indexM :: (Shape f, Storable a, Monad m) => SArray f a -> f Int -> m a

O(1) Indexing in a monad.

The monad allows operations to be strict in the vector when necessary. Suppose vector copying is implemented like this:

copy mv v = ... write mv i (v ! i) ...

For lazy vectors, v ! i would not be evaluated which means that mv would unnecessarily retain a reference to v in each element written.

With indexM, copying can be implemented like this instead:

copy mv v = ... do
  x <- indexM v i
  write mv i x

Here, no references to v are retained because indexing (but not the elements) is evaluated eagerly.

Throws an error if the index is out of range.

unsafeIndexM :: (Shape f, Storable a, Monad m) => SArray f a -> f Int -> m a

O(1) Indexing in a monad without bounds checks. See indexM for an explanation of why this is useful.

linearIndexM :: (Shape f, Storable a, Monad m) => SArray f a -> Int -> m a

O(1) Indexing in a monad. Throws an error if the index is out of range.

unsafeLinearIndexM :: (Storable a, Monad m) => SArray f a -> Int -> m a

O(1) Indexing in a monad without bounds checks. See indexM for an explanation of why this is useful.

Modifying arrays

Bulk updates

(//) :: (Storable a, Shape f) => SArray f a -> [(f Int, a)] -> SArray f a

For each pair (i,a) from the list, replace the array element at position i by a.

Accumulations

accum

Arguments

:: (Shape f, Storable a)

=> (a -> b -> a)
accumulating function f

-> SArray f a
initial array

-> [(f Int, b)]
list of index/value pairs (of length n)

-> SArray f a

O(m+n) For each pair (i,b) from the list, replace the array element a at position i by f a b.

Mapping

map :: (Storable a, Storable b) => (a -> b) -> SArray f a -> SArray f b

O(n) Map a function over an array

imap :: (Shape f, Storable a, Storable b) => (f Int -> a -> b) -> SArray f a -> SArray f b

O(n) Apply a function to every element of a vector and its index

Zipping

Zip with function

zipWith :: (Shape f, Storable a, Storable b, Storable c) => (a -> b -> c) -> SArray f a -> SArray f b -> SArray f c

Zip two arrays using the given function. If the array's don't have the same shape, the new array with be the intersection of the two shapes.

zipWith3 :: (Shape f, Storable a, Storable b, Storable c, Storable d) => (a -> b -> c -> d) -> SArray f a -> SArray f b -> SArray f c -> SArray f d

Zip three arrays using the given function. If the array's don't have the same shape, the new array with be the intersection of the two shapes.

izipWith :: (Shape f, Storable a, Storable b, Storable c) => (f Int -> a -> b -> c) -> SArray f a -> SArray f b -> SArray f c

Zip two arrays using the given function with access to the index. If the array's don't have the same shape, the new array with be the intersection of the two shapes.

izipWith3 :: (Shape f, Storable a, Storable b, Storable c, Storable d) => (f Int -> a -> b -> c -> d) -> SArray f a -> SArray f b -> SArray f c -> SArray f d

Zip two arrays using the given function with access to the index. If the array's don't have the same shape, the new array with be the intersection of the two shapes.

Slices

Matrix

ixRow :: Storable a => Int -> IndexedTraversal' Int (SArray V2 a) (Vector a)

Affine traversal over a single row in a matrix.

>>> traverseOf_ rows print $ m & ixRow 1 . each +~ 2
[1,2,3]
[6,7,8]

The row vector should remain the same size to satisfy traversal laws but give reasonable behaviour if the size differs:

>>> traverseOf_ rows print $ m & ixRow 1 .~ V.fromList [0,1]
[1,2,3]
[0,1,6]

>>> traverseOf_ rows print $ m & ixRow 1 .~ V.fromList [0..100]
[1,2,3]
[0,1,2]

rows :: (Storable a, Storable b) => IndexedTraversal Int (SArray V2 a) (SArray V2 b) (Vector a) (Vector b)

Indexed traversal over the rows of a matrix. Each row is an efficient slice of the original vector.

>>> traverseOf_ rows print m
[1,2,3]
[4,5,6]

ixColumn :: Storable a => Int -> IndexedTraversal' Int (SArray V2 a) (Vector a)

Affine traversal over a single column in a matrix.

>>> traverseOf_ rows print $ m & ixColumn 2 . each *~ 10
[1,2,30]
[4,5,60]

columns :: (Storable a, Storable b) => IndexedTraversal Int (SArray V2 a) (SArray V2 b) (Vector a) (Vector b)

Indexed traversal over the columns of a matrix. Unlike rows, each column is a new separate vector.

>>> traverseOf_ columns print m
[1,4]
[2,5]
[3,6]

>>> traverseOf_ rows print $ m & columns . indices odd . each .~ 0
[1,0,3]
[4,0,6]

The vectors should be the same size to be a valid traversal. If the vectors are different sizes, the number of rows in the new array will be the length of the smallest vector.

3D

ixPlane :: Storable a => ALens' (V3 Int) (V2 Int) -> Int -> IndexedTraversal' Int (SArray V3 a) (SArray V2 a)

Traversal over a single plane of a 3D array given a lens onto that plane (like _xy, _yz, _zx).

planes :: (Storable a, Storable b) => ALens' (V3 Int) (V2 Int) -> IndexedTraversal Int (SArray V3 a) (SArray V3 b) (SArray V2 a) (SArray V2 b)

Traversal over all planes of 3D array given a lens onto that plane (like _xy, _yz, _zx).

flattenPlane :: (Storable a, Storable b) => ALens' (V3 Int) (V2 Int) -> (Vector a -> b) -> SArray V3 a -> SArray V2 b

Flatten a plane by reducing a vector in the third dimension to a single value.

Ordinals

unsafeOrdinals :: (Storable a, Shape f) => [f Int] -> IndexedTraversal' (f Int) (SArray f a) a

This Traversal should not have any duplicates in the list of indices.

Mutable

type SMArray = MArray MVector

Storable mutable array.

thaw :: (PrimMonad m, Storable a) => SArray f a -> m (SMArray f (PrimState m) a)

O(n) Yield an immutable copy of the mutable array.

freeze :: (PrimMonad m, Storable a) => SMArray f (PrimState m) a -> m (SArray f a)

O(n) Yield a mutable copy of the immutable vector.

unsafeThaw :: (PrimMonad m, Storable a) => SArray f a -> m (SMArray f (PrimState m) a)

O(1) Unsafely convert an immutable array to a mutable one without copying. The immutable array may not be used after this operation.

unsafeFreeze :: (PrimMonad m, Storable a) => SMArray f (PrimState m) a -> m (SArray f a)

O(1) Unsafe convert a mutable array to an immutable one without copying. The mutable array may not be used after this operation.

Delayed

data Delayed f a

A delayed representation of an array. This useful for mapping over an array in parallel.

Generating delayed

delayed :: (Storable a, Storable b, Shape f, Shape k) => Iso (SArray f a) (SArray k b) (Delayed f a) (Delayed k b)

Isomorphism between an array and its delayed representation. Conversion to the array is done in parallel.

seqDelayed :: (Storable a, Storable b, Shape f, Shape k) => Iso (SArray f a) (SArray k b) (Delayed f a) (Delayed k b)

Isomorphism between an array and its delayed representation. Conversion to the array is done in sequence.

delay :: (Storable a, Shape f) => SArray f a -> Delayed f a

Turn a material array into a delayed one with the same shape.

manifest :: (Storable a, Shape f) => Delayed f a -> SArray f a

Parallel manifestation of a delayed array into a material one.

seqManifest :: (Storable a, Shape f) => Delayed f a -> SArray f a

Sequential manifestation of a delayed array.

genDelayed :: Layout f -> (f Int -> a) -> Delayed f a

Generate a Delayed array using the given Layout and construction function.

indexDelayed :: Shape f => Delayed f a -> f Int -> a

Index a delayed array, returning a IndexOutOfBounds exception if the index is out of range.

affirm :: (Shape f, Storable a) => Delayed f a -> Delayed f a

manifest an array to a SArray and delay again.

seqAffirm :: (Shape f, Storable a) => Delayed f a -> Delayed f a

seqManifest an array to a SArray and delay again.

Focused

data Focused f a

A delayed representation of an array with a focus on a single element. This element is the target of extract.

Generating focused

focusOn :: f Int -> Delayed f a -> Focused f a

Focus on a particular element of a delayed array.

unfocus :: Focused f a -> Delayed f a

Discard the focus to retrieve the delayed array.

unfocused :: IndexedLens (f Int) (Focused f a) (Focused f b) (Delayed f a) (Delayed f b)

Indexed lens onto the delayed array, indexed at the focus.

extendFocus :: Shape f => (Focused f a -> b) -> Delayed f a -> Delayed f b

Modify a Delayed array by extracting a value from a Focused each point.

Focus location

locale :: ComonadStore s w => Lens' (w a) s

Lens onto the position of a ComonadStore.

locale :: Lens' (Focused l a) (l Int)

shiftFocus :: Applicative f => f Int -> Focused f a -> Focused f a

Focus on a neighbouring element, relative to the current focus.

Pointers

unsafeWithPtr :: Storable a => SArray f a -> (Ptr a -> IO b) -> IO b

Pass a pointer to the array's data to the IO action. Modifying data through the Ptr is unsafe.

unsafeToForeignPtr :: Storable a => SArray f a -> ForeignPtr a

Yield the underlying ForeignPtr. Modifying the data through the ForeignPtr is unsafe.

unsafeFromForeignPtr :: (Shape f, Storable a) => Layout f -> ForeignPtr a -> SArray f a

O(1) Create an array from a layout and ForeignPtr. It is assumed the pointer points directly to the data (no offset). Modifying data through the ForeignPtr afterwards is unsafe.